Revert "Alerting: Enable alerts preview on notification policies page (#68291)"
This reverts commit f7b8a666f4ce78e6f44fe78148e605dcc515ab14.
