+++ public/app/features/alerting/unified/api/alertmanagerApi.ts
@@ -2,7 +2,6 @@ import {
   AlertmanagerAlert,
   AlertmanagerChoice,
   AlertManagerCortexConfig,
-  AlertmanagerGroup,
   ExternalAlertmanagerConfig,
   ExternalAlertmanagers,
   ExternalAlertmanagersResponse,
@@ -60,12 +59,6 @@ export const alertmanagerApi = alertingApi.injectEndpoints({
           params,
         };
       },
-    }),
-
-    getAlertmanagerAlertGroups: build.query<AlertmanagerGroup[], { amSourceName: string }>({
-      query: ({ amSourceName }) => ({
-        url: `/api/alertmanager/${getDatasourceAPIUid(amSourceName)}/api/v2/alerts/groups`,
-      }),
     }),
 
     getAlertmanagerChoiceStatus: build.query<AlertmanagersChoiceResponse, void>({
