+++ public/app/features/alerting/unified/features.ts
@@ -1,7 +1,14 @@
 import { FeatureDescription } from 'react-enable/dist/FeatureState';
 
-export enum AlertingFeature {}
+export enum AlertingFeature {
+  NotificationPoliciesV2MatchingInstances = 'notification-policies.v2.matching-instances',
+}
 
-const FEATURES: FeatureDescription[] = [];
+const FEATURES: FeatureDescription[] = [
+  {
+    name: AlertingFeature.NotificationPoliciesV2MatchingInstances,
+    defaultValue: false,
+  },
+];
 
 export default FEATURES;
