+++ public/app/features/alerting/unified/utils/alertmanager.ts
@@ -126,6 +126,41 @@ export const matcherFieldOptions: SelectableValue[] = [
   { label: MatcherOperator.notRegex, description: 'Does not match regex', value: MatcherOperator.notRegex },
 ];
 
+const matcherOperators = [
+  MatcherOperator.regex,
+  MatcherOperator.notRegex,
+  MatcherOperator.notEqual,
+  MatcherOperator.equal,
+];
+
+export function parseMatcher(matcher: string): Matcher {
+  const trimmed = matcher.trim();
+  if (trimmed.startsWith('{') && trimmed.endsWith('}')) {
+    throw new Error(`PromQL matchers not supported yet, sorry! PromQL matcher found: ${trimmed}`);
+  }
+  const operatorsFound = matcherOperators
+    .map((op): [MatcherOperator, number] => [op, trimmed.indexOf(op)])
+    .filter(([_, idx]) => idx > -1)
+    .sort((a, b) => a[1] - b[1]);
+
+  if (!operatorsFound.length) {
+    throw new Error(`Invalid matcher: ${trimmed}`);
+  }
+  const [operator, idx] = operatorsFound[0];
+  const name = trimmed.slice(0, idx).trim();
+  const value = trimmed.slice(idx + operator.length).trim();
+  if (!name) {
+    throw new Error(`Invalid matcher: ${trimmed}`);
+  }
+
+  return {
+    name,
+    value,
+    isRegex: operator === MatcherOperator.regex || operator === MatcherOperator.notRegex,
+    isEqual: operator === MatcherOperator.equal || operator === MatcherOperator.regex,
+  };
+}
+
 export function matcherToObjectMatcher(matcher: Matcher): ObjectMatcher {
   const operator = matcherToOperator(matcher);
   return [matcher.name, operator, matcher.value];
