+++ public/app/features/alerting/unified/utils/amroutes.test.ts
@@ -1,8 +1,8 @@
-import { Route } from 'app/plugins/datasource/alertmanager/types';
+import { MatcherOperator, Route } from 'app/plugins/datasource/alertmanager/types';
 
 import { FormAmRoute } from '../types/amroutes';
 
-import { amRouteToFormAmRoute, emptyRoute, formAmRouteToAmRoute } from './amroutes';
+import { amRouteToFormAmRoute, emptyRoute, formAmRouteToAmRoute, normalizeMatchers } from './amroutes';
 
 const emptyAmRoute: Route = {
   receiver: '',
@@ -87,5 +87,30 @@ describe('amRouteToFormAmRoute', () => {
       expect(formRoute.groupBy).toStrictEqual(['SHOULD BE SET']);
       expect(formRoute.overrideGrouping).toBe(true);
     });
+  });
+});
+
+describe('normalizeMatchers', () => {
+  const eq = MatcherOperator.equal;
+
+  it('should work for object_matchers', () => {
+    const route: Route = { object_matchers: [['foo', eq, 'bar']] };
+    expect(normalizeMatchers(route)).toEqual([['foo', eq, 'bar']]);
+  });
+  it('should work for matchers', () => {
+    const route: Route = { matchers: ['foo=bar', 'foo!=bar', 'foo=~bar', 'foo!~bar'] };
+    expect(normalizeMatchers(route)).toEqual([
+      ['foo', MatcherOperator.equal, 'bar'],
+      ['foo', MatcherOperator.notEqual, 'bar'],
+      ['foo', MatcherOperator.regex, 'bar'],
+      ['foo', MatcherOperator.notRegex, 'bar'],
+    ]);
+  });
+  it('should work for match and match_re', () => {
+    const route: Route = { match: { foo: 'bar' }, match_re: { foo: 'bar' } };
+    expect(normalizeMatchers(route)).toEqual([
+      ['foo', MatcherOperator.regex, 'bar'],
+      ['foo', MatcherOperator.equal, 'bar'],
+    ]);
   });
 });
