+++ public/app/features/alerting/unified/utils/notification-policies.ts
@@ -1,25 +1,61 @@
-import { AlertmanagerGroup, Route, RouteWithID } from 'app/plugins/datasource/alertmanager/types';
+import { AlertmanagerGroup, MatcherOperator, ObjectMatcher, Route } from 'app/plugins/datasource/alertmanager/types';
 
-import { Label, normalizeMatchers, labelsMatchObjectMatchers } from './matchers';
+import { normalizeMatchers } from './amroutes';
+
+export type Label = [string, string];
+type OperatorPredicate = (labelValue: string, matcherValue: string) => boolean;
+
+const OperatorFunctions: Record<MatcherOperator, OperatorPredicate> = {
+  [MatcherOperator.equal]: (lv, mv) => lv === mv,
+  [MatcherOperator.notEqual]: (lv, mv) => lv !== mv,
+  [MatcherOperator.regex]: (lv, mv) => Boolean(lv.match(new RegExp(mv))),
+  [MatcherOperator.notRegex]: (lv, mv) => !Boolean(lv.match(new RegExp(mv))),
+};
+
+function isLabelMatch(matcher: ObjectMatcher, label: Label) {
+  const [labelKey, labelValue] = label;
+  const [matcherKey, operator, matcherValue] = matcher;
+
+  // not interested, keys don't match
+  if (labelKey !== matcherKey) {
+    return false;
+  }
+
+  const matchFunction = OperatorFunctions[operator];
+  if (!matchFunction) {
+    throw new Error(`no such operator: ${operator}`);
+  }
+
+  return matchFunction(labelValue, matcherValue);
+}
+
+// check if every matcher returns "true" for the set of labels
+function matchLabels(matchers: ObjectMatcher[], labels: Label[]) {
+  return matchers.every((matcher) => {
+    return labels.some((label) => isLabelMatch(matcher, label));
+  });
+}
 
 // Match does a depth-first left-to-right search through the route tree
 // and returns the matching routing nodes.
-function findMatchingRoutes(root: Route, labels: Label[]): Route[] {
-  const matches: Route[] = [];
+function findMatchingRoutes<T extends Route>(root: T, labels: Label[]): T[] {
+  let matches: T[] = [];
 
   // If the current node is not a match, return nothing
-  // const normalizedMatchers = normalizeMatchers(root);
-  // Normalization should have happened earlier in the code
-  if (!root.object_matchers || !labelsMatchObjectMatchers(root.object_matchers, labels)) {
+  const normalizedMatchers = normalizeMatchers(root);
+  if (!matchLabels(normalizedMatchers, labels)) {
     return [];
   }
 
   // If the current node matches, recurse through child nodes
   if (root.routes) {
-    for (const child of root.routes) {
-      const matchingChildren = findMatchingRoutes(child, labels);
+    for (let index = 0; index < root.routes.length; index++) {
+      let child = root.routes[index];
+      let matchingChildren = findMatchingRoutes(child, labels);
 
-      matches.push(...matchingChildren);
+      // TODO how do I solve this typescript thingy? It looks correct to me /shrug
+      // @ts-ignore
+      matches = matches.concat(matchingChildren);
 
       // we have matching children and we don't want to continue, so break here
       if (matchingChildren.length && !child.continue) {
@@ -34,22 +70,6 @@ function findMatchingRoutes(root: Route, labels: Label[]): Route[] {
   }
 
   return matches;
-}
-
-// This is a performance improvement to normalize matchers only once and use the normalized version later on
-export function normalizeRoute(rootRoute: RouteWithID): RouteWithID {
-  function normalizeRoute(route: RouteWithID) {
-    route.object_matchers = normalizeMatchers(route);
-    delete route.matchers;
-    delete route.match;
-    delete route.match_re;
-    route.routes?.forEach(normalizeRoute);
-  }
-
-  const normalizedRootRoute = structuredClone(rootRoute);
-  normalizeRoute(normalizedRootRoute);
-
-  return normalizedRootRoute;
 }
 
 /**
@@ -82,4 +102,4 @@ function findMatchingAlertGroups(
   }, matchingGroups);
 }
 
-export { findMatchingAlertGroups, findMatchingRoutes };
+export { findMatchingAlertGroups, findMatchingRoutes, matchLabels };
